MVP-5076-1.5: Refactor Connect.tsx/ Add new hook useConnect
