+++ packages/notifi-react/lib/utils/constants.ts
@@ -8,7 +8,7 @@ export const defaultCopy = {
   connect: {
     title: 'Notifications',
     description: 'See your activity and gain access to alerts',
-    hardwareWalletLabel: 'Use Hardware WalletList',
+    hardwareWalletLabel: 'Use Hardware Wallet',
     buttonText: 'Next',
   },
   ftuAlertList: {
